initial commit
What I currently have so far. Can get info from the game and display it
on a JPanel called telemetry. Not the prettiest looking at the moment
but it's more of a proof of concept and will be used till I finialize
the design then will be looking at sending the telemetry info to other
programs